requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-decouple==3.8
PyJWT==2.8.1
Pillow==10.1.0
